Extracted preference api to its own
