+++ e2e/cypress/support/api_commands.d.ts
@@ -188,17 +188,5 @@ declare namespace Cypress {
          *   });
          */
         apiAddUserToChannel(channelId: string, userId: string): Chainable<Response>;
-
-        /**
-         * Save a list of the user's preferences.
-         * See https://api.mattermost.com/#tag/preferences/paths/~1users~1{user_id}~1preferences/put
-         * @param {object[]} preferences - List of preference objects
-         * @param {string} userId - User ID
-         * @returns {Response} response: Cypress-chainable response which should have successful HTTP status of 200 OK to continue or pass.
-         *
-         * @example
-         *   cy.apiSaveUserPreference([{user_id: 'user-id', category: 'display_settings', name: 'channel_display_mode', value: 'full'}], 'user-id');
-         */
-        apiSaveUserPreference(preferences: Array<object>, userId: string): Chainable<Response>;
     }
 }
